[Presentation] Move Extensions
